clean: Separate styled components on large files
