+++ src/sidebar/legal-links/legal-links.tsx
@@ -1,4 +1,5 @@
 import { getLegalLinks } from '../../services/static-data';
+import { LegalLink, LegalLinksList, LegalLinksWrapper } from './styles';
 
 const legalLinks = getLegalLinks();
